Removing bigtable-hbase/TestBigtableConnection I'm moving GetTableMicrobenchmark and PutMicroBenchmark into hbase-1.x so that we can use the default implementation.
